Split c2c
